Add tests for device, device_manager, and wallet_manager
