+++ src/cryptoadvance/specter/templates/includes/sidebar/sidebar.jinja
@@ -34,7 +34,7 @@
             Devices
         </div>
         {% from 'includes/sidebar/components/sidebar_device_list_item.jinja' import sidebar_device_list_item %}
-        {% for device_name in specter.device_manager.devices %}
+        {% for device_name in specter.device_manager.devices_names %}
                 {{ sidebar_device_list_item(specter.device_manager.devices[device_name], device_alias) }}
         {% endfor %}
         {{ sidebar_btn('new_device', 'Add new device') }}
